Разбивает компонент попапа на сложный компонент с под-компонентами
